+++ AMS/Models/AdvocateSchedule.cs
@@ -2,11 +2,9 @@
 
 public class AdvocateSchedule : BaseEntity
 {
-   
-
     public DateTime ScheduleDate { get; set; }
     public string AvailableTime { get; set; }
 
-    public int AdvocateId { get; set; }
+    public long AdvocateId { get; set; }
     public Advocate Advocate { get; set; }
 }
